Complete reworking of tests, part 3
