+++ hendrics/tests/test_a_complete_run.py
@@ -167,291 +167,6 @@ class TestFullRun(object):
         gti_to_test = hen.io.load_events(self.ev_fileA).gti
         assert np.allclose(gti_to_test, out_lc.gti)
 
-    def test_pds_leahy_dtbig(self):
-        """Test PDS production."""
-        lc = self.lcA
-        hen.io.main([lc])
-        command = \
-            '{0} -f 128 -k PDS --save-all --norm leahy -b {1}'.format(lc, 1)
-        hen.fspec.main(command.split())
-
-        assert os.path.exists(
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds' + HEN_FILE_EXTENSION))
-
-    def test_pds_leahy(self):
-        """Test PDS production."""
-        lc = os.path.join(self.datadir,
-                          'monol_testA_E3-50_lc') + HEN_FILE_EXTENSION
-        hen.io.main([lc])
-        command = \
-            '{0} -f 128 -k PDS --save-all --norm leahy'.format(lc)
-        hen.fspec.main(command.split())
-
-        assert os.path.exists(
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds' + HEN_FILE_EXTENSION))
-
-    def test_pds(self):
-        """Test PDS production."""
-        command = \
-            '{0} {1} -f 128 --save-all --save-dyn -k PDS ' \
-            '--norm frac'.format(
-                os.path.join(self.datadir,
-                             'monol_testA_E3-50_lc') + HEN_FILE_EXTENSION,
-                os.path.join(self.datadir,
-                             'monol_testB_E3-50_lc') + HEN_FILE_EXTENSION)
-        hen.fspec.main(command.split())
-
-        assert os.path.exists(
-            os.path.join(self.datadir,
-                         'monol_testB_E3-50_pds' + HEN_FILE_EXTENSION))
-        assert os.path.exists(
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds') + HEN_FILE_EXTENSION)
-
-    def test_cpds_rms_norm(self):
-        """Test CPDS production."""
-        command = \
-            '{0} {1} -f 128 --save-dyn -k CPDS --save-all ' \
-            '--norm rms -o {2}'.format(
-                os.path.join(self.datadir, 'monol_testA_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_testB_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_test_E3-50'))
-
-        hen.fspec.main(command.split())
-
-    def test_cpds_wrong_norm(self):
-        """Test CPDS production."""
-        command = \
-            '{0} {1} -f 128 --save-dyn -k CPDS --norm blablabla -o {2}'.format(
-                os.path.join(self.datadir, 'monol_testA_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_testB_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_test_E3-50'))
-        with pytest.warns(UserWarning) as record:
-            hen.fspec.main(command.split())
-
-        assert np.any(["Beware! Unknown normalization" in r.message.args[0]
-                       for r in record])
-
-    def test_cpds_dtbig(self):
-        """Test CPDS production."""
-        command = \
-            '{0} {1} -f 128 --save-dyn -k CPDS --save-all --norm ' \
-            'frac -o {2}'.format(
-                os.path.join(self.datadir, 'monol_testA_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_testB_E3-50_lc') +
-                HEN_FILE_EXTENSION,
-                os.path.join(self.datadir, 'monol_test_E3-50'))
-        command += ' -b 1'
-        hen.fspec.main(command.split())
-
-    def test_dumpdynpds(self):
-        """Test dump dynamical PDSs."""
-        command = '--noplot ' + \
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds') + \
-            HEN_FILE_EXTENSION
-        with pytest.raises(NotImplementedError):
-            hen.fspec.dumpdyn_main(command.split())
-
-    def test_sumpds(self):
-        """Test the sum of pdss."""
-        hen.sum_fspec.main([
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds') + HEN_FILE_EXTENSION,
-            os.path.join(self.datadir,
-                         'monol_testB_E3-50_pds') + HEN_FILE_EXTENSION,
-            '-o', os.path.join(self.datadir,
-                               'monol_test_sum' + HEN_FILE_EXTENSION)])
-
-    def test_dumpdyncpds(self):
-        """Test dump dynamical PDSs."""
-        command = '--noplot ' + \
-            os.path.join(self.datadir,
-                         'monol_test_E3-50_cpds') + \
-        HEN_FILE_EXTENSION
-        with pytest.raises(NotImplementedError):
-            hen.fspec.dumpdyn_main(command.split())
-
-    def test_rebinpds(self):
-        """Test PDS rebinning 1."""
-        command = '{0} -r 2'.format(
-            os.path.join(self.datadir, 'monol_testA_E3-50_pds') +
-            HEN_FILE_EXTENSION)
-        hen.rebin.main(command.split())
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_testA_E3-50_pds_rebin2' +
-                                    HEN_FILE_EXTENSION))
-
-    def test_rebinpds_geom(self):
-        """Test geometrical PDS rebinning."""
-        command = '{0} {1} -r 1.03'.format(
-            os.path.join(self.datadir, 'monol_testA_E3-50_pds') +
-            HEN_FILE_EXTENSION,
-            os.path.join(self.datadir, 'monol_testB_E3-50_pds') +
-            HEN_FILE_EXTENSION
-            )
-        hen.rebin.main(command.split())
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_testA_E3-50_pds_rebin1.03' +
-                                    HEN_FILE_EXTENSION))
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_testB_E3-50_pds_rebin1.03' +
-                                    HEN_FILE_EXTENSION))
-
-    def test_rebincpds(self):
-        """Test CPDS rebinning."""
-        command = '{0} -r 2'.format(
-            os.path.join(self.datadir, 'monol_test_E3-50_cpds') +
-            HEN_FILE_EXTENSION)
-        hen.rebin.main(command.split())
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_test_E3-50_cpds_rebin2' +
-                                    HEN_FILE_EXTENSION))
-
-    def test_rebincpds_geom(self):
-        """Test CPDS geometrical rebinning."""
-        command = '{0} -r 1.03'.format(
-            os.path.join(self.datadir, 'monol_test_E3-50_cpds') +
-            HEN_FILE_EXTENSION)
-        hen.rebin.main(command.split())
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_test_E3-50_cpds_rebin1.03' +
-                                    HEN_FILE_EXTENSION))
-
-    def test_save_lags(self):
-        fname = os.path.join(self.datadir,
-                             'monol_test_E3-50_cpds_rebin2' +
-                             HEN_FILE_EXTENSION)
-        hen.timelags.main([fname])
-        out = hen.base.hen_root(fname) + '_lags.qdp'
-        os.path.exists(out)
-
-    def test_fit_pds(self):
-        modelstring = '''
-from astropy.modeling import models
-model = models.Const1D()
-        '''
-        modelfile = 'bubu__model__.py'
-        with open(modelfile, 'w') as fobj:
-            print(modelstring, file=fobj)
-        pdsfile1 = \
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds' + HEN_FILE_EXTENSION)
-        pdsfile2 = \
-            os.path.join(self.datadir,
-                         'monol_testB_E3-50_pds' + HEN_FILE_EXTENSION)
-
-        command = '{0} {1} -m {2} --frequency-interval 0 10'.format(
-            pdsfile1,
-            pdsfile2,
-            modelfile)
-        hen.modeling.main_model(command.split())
-
-        out0 = os.path.join(self.datadir,
-                            'monol_testA_E3-50_pds_bestfit.p')
-        out1 = os.path.join(self.datadir,
-                            'monol_testB_E3-50_pds_bestfit.p')
-        assert os.path.exists(out0)
-        assert os.path.exists(out1)
-        m, k, c = hen.io.load_model(
-            os.path.join(self.datadir,
-                         'monol_testB_E3-50_pds_bestfit.p'))
-        assert hasattr(m, 'amplitude')
-        os.unlink(out0)
-        os.unlink(out1)
-
-        out0 = os.path.join(self.datadir,
-                            'monol_testA_E3-50_pds_fit' + HEN_FILE_EXTENSION)
-        out1 = os.path.join(self.datadir,
-                            'monol_testB_E3-50_pds_fit' + HEN_FILE_EXTENSION)
-        assert os.path.exists(out0)
-        assert os.path.exists(out1)
-        spec = hen.io.load_pds(out0)
-        assert hasattr(spec, 'best_fits')
-
-    def test_fit_cpds(self):
-        modelstring = '''
-from astropy.modeling import models
-model = models.Const1D()
-        '''
-        modelfile = 'bubu__model__.py'
-        with open(modelfile, 'w') as fobj:
-            print(modelstring, file=fobj)
-        pdsfile1 = \
-            os.path.join(self.datadir,
-                         'monol_test_E3-50_cpds' + HEN_FILE_EXTENSION)
-
-        command = '{0} -m {1} --frequency-interval 0 10'.format(
-            pdsfile1,
-            modelfile)
-        hen.modeling.main_model(command.split())
-
-        out0 = os.path.join(self.datadir,
-                            'monol_test_E3-50_cpds_bestfit.p')
-        assert os.path.exists(out0)
-        m, k, c = hen.io.load_model(out0)
-        assert hasattr(m, 'amplitude')
-        os.unlink(out0)
-
-        out0 = \
-            os.path.join(self.datadir,
-                         'monol_test_E3-50_cpds_fit' + HEN_FILE_EXTENSION)
-        assert os.path.exists(out0)
-        spec = hen.io.load_pds(out0)
-        assert hasattr(spec, 'best_fits')
-
-    def test_fit_pds_f_no_of_intervals_invalid(self):
-        modelstring = '''
-from astropy.modeling import models
-model = models.Const1D()
-        '''
-        modelfile = 'bubu__model__.py'
-        with open(modelfile, 'w') as fobj:
-            print(modelstring, file=fobj)
-        pdsfile1 = \
-            os.path.join(self.datadir,
-                         'monol_testA_E3-50_pds' + HEN_FILE_EXTENSION)
-        pdsfile2 = \
-            os.path.join(self.datadir,
-                         'monol_testB_E3-50_pds' + HEN_FILE_EXTENSION)
-
-        command = '{0} {1} -m {2} --frequency-interval 0 1 9'.format(pdsfile1,
-                                                                     pdsfile2,
-                                                                     modelfile)
-        with pytest.raises(ValueError) as excinfo:
-            hen.modeling.main_model(command.split())
-
-        assert "Invalid number of frequencies specified" in str(excinfo.value)
-
-    def test_savexspec(self):
-        """Test save as Xspec 1."""
-        command = '{0}'.format(
-            os.path.join(self.datadir, 'monol_testA_E3-50_pds_rebin2') +
-            HEN_FILE_EXTENSION)
-        hen.save_as_xspec.main(command.split())
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_testA_E3-50_pds_rebin2.pha'))
-
-    def test_savexspec_geom(self):
-        """Test save as Xspec 2."""
-        command = '{0}'.format(
-            os.path.join(self.datadir, 'monol_test_E3-50_cpds_rebin1.03') +
-            HEN_FILE_EXTENSION)
-        hen.save_as_xspec.main(command.split())
-
-        os.path.exists(os.path.join(self.datadir,
-                                    'monol_test_E3-50_cpds_rebin1.03.pha'))
-        os.path.exists(
-            os.path.join(self.datadir,
-                         'monol_test_E3-50_cpds_rebin1.03_lags.pha'))
-
     def test_readfile_fits(self):
         """Test reading and dumping a FITS file."""
         fitsname = os.path.join(self.datadir, 'monol_testA.evt')
@@ -479,44 +194,6 @@ model = models.Const1D()
             [array, errors],
             filename=os.path.join(self.datadir, "monol_test.txt"),
             colnames=["array", "err"])
-
-    def test_plot_lin(self):
-        """Test plotting with linear axes."""
-        pname = os.path.join(self.datadir, 'monol_testA_E3-50_pds_fit') + \
-            HEN_FILE_EXTENSION
-        cname = os.path.join(self.datadir, 'monol_test_E3-50_cpds_fit') + \
-            HEN_FILE_EXTENSION
-        lname = os.path.join(self.datadir, 'monol_testA_E3-50_lc') + \
-            HEN_FILE_EXTENSION
-        hen.plot.main([pname, cname, lname, '--noplot', '--xlin', '--ylin',
-                      '-o', 'dummy.qdp'])
-        hen.plot.main([lname, '--noplot',
-                       '--axes', 'time', 'counts', '--xlin', '--ylin',
-                       '-o', 'dummy.qdp'])
-
-    def test_plot_log(self):
-        """Test plotting with log axes."""
-        pname = os.path.join(
-            self.datadir,
-            'monol_testA_E3-50_pds_rebin1.03' + HEN_FILE_EXTENSION)
-        cname = os.path.join(
-            self.datadir,
-            'monol_test_E3-50_cpds_rebin1.03' + HEN_FILE_EXTENSION)
-        hen.plot.main([pname, cname, '--noplot', '--xlog', '--ylog',
-                       '-o', 'dummy.qdp'])
-        hen.plot.main([pname, '--noplot', '--axes', 'power', 'power_err',
-                       '--xlin', '--ylin',
-                       '-o', 'dummy.qdp'])
-
-    def test_plot_save_figure(self):
-        """Test plotting and saving figure."""
-        pname = os.path.join(
-            self.datadir,
-            'monol_testA_E3-50_pds_rebin1.03' + HEN_FILE_EXTENSION)
-        hen.plot.main([pname, '--noplot', '--figname',
-                      os.path.join(self.datadir,
-                                   'monol_testA_E3-50_pds_rebin1.03.png'),
-                      '-o', 'dummy.qdp'])
 
     def test_plot_color(self):
         """Test plotting with linear axes."""
