+++ hendrics/tests/test_base.py
@@ -1,0 +1,34 @@
+import pytest
+from astropy.tests.helper import remote_data
+import numpy as np
+from hendrics.base import deorbit_events
+from stingray.events import EventList
+
+
+def test_deorbit_badpar(self):
+    ev = np.asarray(1)
+    with pytest.warns(UserWarning) as record:
+        ev_deor = deorbit_events(ev, None)
+    assert np.any(["No parameter file specified" in r.message.args[0]
+                   for r in record])
+    assert ev_deor == ev
+
+
+def test_deorbit_non_existing_par(self):
+    ev = np.asarray(1)
+    with pytest.raises(FileNotFoundError) as excinfo:
+        ev_deor = deorbit_events(ev, "warjladsfjqpeifjsdk.par")
+    assert "Parameter file warjladsfjqpeifjsdk.par does not exist" \
+           in str(excinfo.value)
+
+
+@remote_data
+@pytest.mark.skipif('not HAS_PINT')
+def test_deorbit_bad_mjdref(self):
+    from hendrics.base import deorbit_events
+    ev = EventList(np.arange(100), gti=np.asarray([[0, 2]]))
+    ev.mjdref = 2
+    par = _dummy_par('bububu.par')
+    with pytest.raises(ValueError) as excinfo:
+        _ = deorbit_events(ev, par)
+    assert "MJDREF is very low (<01-01-1950), " in str(excinfo.value)
